pyproject.toml
[tool.poetry]
name = "meri"
version = "0.2.0"
description = ""
authors = [
    "Christian Möller <christian.moller@novia.fi>",
    "Lamin Jatta <lamin.jatta@novia.fi>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
jupyter = "^1.1.1"
docling = "2.5.2"
docling-parse = "2.0.4"
llama-index = "^0.12.0"
litellm = "^1.52.10"
ipykernel = "^6.29.5"
llama-index-node-parser-docling = "^0.3.0"
llama-index-readers-docling = "^0.3.0"
llama-index-embeddings-huggingface = "^0.4.0"
llama-index-llms-huggingface-api = "^0.3.0"
llama-index-readers-file = "^0.4.0"
python-dotenv = "^1.0.1"
llama-index-llms-litellm = "^0.3.0"
pymupdf = "^1.24.14"
python-fasthtml = ">=0.12.0"
aiofiles = "^24.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"